Extract JS snippets to embedded *.js files
