+++ compiler/prelude/goroutines.js
@@ -1,0 +1,389 @@
+var $stackDepthOffset = 0;
+var $getStackDepth = function () {
+    var err = new Error();
+    if (err.stack === undefined) {
+        return undefined;
+    }
+    return $stackDepthOffset + err.stack.split("\n").length;
+};
+
+var $panicStackDepth = null, $panicValue;
+var $callDeferred = function (deferred, jsErr, fromPanic) {
+    if (!fromPanic && deferred !== null && $curGoroutine.deferStack.indexOf(deferred) == -1) {
+        throw jsErr;
+    }
+    if (jsErr !== null) {
+        var newErr = null;
+        try {
+            $panic(new $jsErrorPtr(jsErr));
+        } catch (err) {
+            newErr = err;
+        }
+        $callDeferred(deferred, newErr);
+        return;
+    }
+    if ($curGoroutine.asleep) {
+        return;
+    }
+
+    $stackDepthOffset--;
+    var outerPanicStackDepth = $panicStackDepth;
+    var outerPanicValue = $panicValue;
+
+    var localPanicValue = $curGoroutine.panicStack.pop();
+    if (localPanicValue !== undefined) {
+        $panicStackDepth = $getStackDepth();
+        $panicValue = localPanicValue;
+    }
+
+    try {
+        while (true) {
+            if (deferred === null) {
+                deferred = $curGoroutine.deferStack[$curGoroutine.deferStack.length - 1];
+                if (deferred === undefined) {
+                    /* The panic reached the top of the stack. Clear it and throw it as a JavaScript error. */
+                    $panicStackDepth = null;
+                    if (localPanicValue.Object instanceof Error) {
+                        throw localPanicValue.Object;
+                    }
+                    var msg;
+                    if (localPanicValue.constructor === $String) {
+                        msg = localPanicValue.$val;
+                    } else if (localPanicValue.Error !== undefined) {
+                        msg = localPanicValue.Error();
+                    } else if (localPanicValue.String !== undefined) {
+                        msg = localPanicValue.String();
+                    } else {
+                        msg = localPanicValue;
+                    }
+                    throw new Error(msg);
+                }
+            }
+            var call = deferred.pop();
+            if (call === undefined) {
+                $curGoroutine.deferStack.pop();
+                if (localPanicValue !== undefined) {
+                    deferred = null;
+                    continue;
+                }
+                return;
+            }
+            var r = call[0].apply(call[2], call[1]);
+            if (r && r.$blk !== undefined) {
+                deferred.push([r.$blk, [], r]);
+                if (fromPanic) {
+                    throw null;
+                }
+                return;
+            }
+
+            if (localPanicValue !== undefined && $panicStackDepth === null) {
+                /* error was recovered */
+                if (fromPanic) {
+                    throw null;
+                }
+                return;
+            }
+        }
+    } catch (e) {
+        // Deferred function threw a JavaScript exception or tries to unwind stack
+        // to the point where a panic was handled.
+        if (fromPanic) {
+            // Re-throw the exception to reach deferral execution call at the end
+            // of the function.
+            throw e;
+        }
+        // We are at the end of the function, handle the error or re-throw to
+        // continue unwinding if necessary, or simply stop unwinding if we got far
+        // enough.
+        $callDeferred(deferred, e, fromPanic);
+    } finally {
+        if (localPanicValue !== undefined) {
+            if ($panicStackDepth !== null) {
+                $curGoroutine.panicStack.push(localPanicValue);
+            }
+            $panicStackDepth = outerPanicStackDepth;
+            $panicValue = outerPanicValue;
+        }
+        $stackDepthOffset++;
+    }
+};
+
+var $panic = function (value) {
+    $curGoroutine.panicStack.push(value);
+    $callDeferred(null, null, true);
+};
+var $recover = function () {
+    if ($panicStackDepth === null || ($panicStackDepth !== undefined && $panicStackDepth !== $getStackDepth() - 2)) {
+        return $ifaceNil;
+    }
+    $panicStackDepth = null;
+    return $panicValue;
+};
+var $throw = function (err) { throw err; };
+
+var $noGoroutine = { asleep: false, exit: false, deferStack: [], panicStack: [] };
+var $curGoroutine = $noGoroutine, $totalGoroutines = 0, $awakeGoroutines = 0, $checkForDeadlock = true, $exportedFunctions = 0;
+var $mainFinished = false;
+var $go = function (fun, args) {
+    $totalGoroutines++;
+    $awakeGoroutines++;
+    var $goroutine = function () {
+        try {
+            $curGoroutine = $goroutine;
+            var r = fun.apply(undefined, args);
+            if (r && r.$blk !== undefined) {
+                fun = function () { return r.$blk(); };
+                args = [];
+                return;
+            }
+            $goroutine.exit = true;
+        } catch (err) {
+            if (!$goroutine.exit) {
+                throw err;
+            }
+        } finally {
+            $curGoroutine = $noGoroutine;
+            if ($goroutine.exit) { /* also set by runtime.Goexit() */
+                $totalGoroutines--;
+                $goroutine.asleep = true;
+            }
+            if ($goroutine.asleep) {
+                $awakeGoroutines--;
+                if (!$mainFinished && $awakeGoroutines === 0 && $checkForDeadlock && $exportedFunctions === 0) {
+                    console.error("fatal error: all goroutines are asleep - deadlock!");
+                    if ($global.process !== undefined) {
+                        $global.process.exit(2);
+                    }
+                }
+            }
+        }
+    };
+    $goroutine.asleep = false;
+    $goroutine.exit = false;
+    $goroutine.deferStack = [];
+    $goroutine.panicStack = [];
+    $schedule($goroutine);
+};
+
+var $scheduled = [];
+var $runScheduled = function () {
+    // For nested setTimeout calls browsers enforce 4ms minimum delay. We minimize
+    // the effect of this penalty by queueing the timer preemptively before we run
+    // the goroutines, and later cancelling it if it turns out unneeded. See:
+    // https://developer.mozilla.org/en-US/docs/Web/API/setTimeout#nested_timeouts
+    var nextRun = setTimeout($runScheduled);
+    try {
+        var start = Date.now();
+        var r;
+        while ((r = $scheduled.shift()) !== undefined) {
+            r();
+            // We need to interrupt this loop in order to allow the event loop to
+            // process timers, IO, etc. However, invoking scheduling through
+            // setTimeout is ~1000 times more expensive, so we amortize this cost by
+            // looping until the 4ms minimal delay has elapsed (assuming there are
+            // scheduled goroutines to run), and then yield to the event loop.
+            var elapsed = Date.now() - start;
+            if (elapsed > 4 || elapsed < 0) { break; }
+        }
+    } finally {
+        if ($scheduled.length == 0) {
+            // Cancel scheduling pass if there's nothing to run.
+            clearTimeout(nextRun);
+        }
+    }
+};
+
+var $schedule = function (goroutine) {
+    if (goroutine.asleep) {
+        goroutine.asleep = false;
+        $awakeGoroutines++;
+    }
+    $scheduled.push(goroutine);
+    if ($curGoroutine === $noGoroutine) {
+        $runScheduled();
+    }
+};
+
+var $setTimeout = function (f, t) {
+    $awakeGoroutines++;
+    return setTimeout(function () {
+        $awakeGoroutines--;
+        f();
+    }, t);
+};
+
+var $block = function () {
+    if ($curGoroutine === $noGoroutine) {
+        $throwRuntimeError("cannot block in JavaScript callback, fix by wrapping code in goroutine");
+    }
+    $curGoroutine.asleep = true;
+};
+
+var $restore = function (context, params) {
+    if (context !== undefined && context.$blk !== undefined) {
+        return context;
+    }
+    return params;
+}
+
+var $send = function (chan, value) {
+    if (chan.$closed) {
+        $throwRuntimeError("send on closed channel");
+    }
+    var queuedRecv = chan.$recvQueue.shift();
+    if (queuedRecv !== undefined) {
+        queuedRecv([value, true]);
+        return;
+    }
+    if (chan.$buffer.length < chan.$capacity) {
+        chan.$buffer.push(value);
+        return;
+    }
+
+    var thisGoroutine = $curGoroutine;
+    var closedDuringSend;
+    chan.$sendQueue.push(function (closed) {
+        closedDuringSend = closed;
+        $schedule(thisGoroutine);
+        return value;
+    });
+    $block();
+    return {
+        $blk: function () {
+            if (closedDuringSend) {
+                $throwRuntimeError("send on closed channel");
+            }
+        }
+    };
+};
+var $recv = function (chan) {
+    var queuedSend = chan.$sendQueue.shift();
+    if (queuedSend !== undefined) {
+        chan.$buffer.push(queuedSend(false));
+    }
+    var bufferedValue = chan.$buffer.shift();
+    if (bufferedValue !== undefined) {
+        return [bufferedValue, true];
+    }
+    if (chan.$closed) {
+        return [chan.$elem.zero(), false];
+    }
+
+    var thisGoroutine = $curGoroutine;
+    var f = { $blk: function () { return this.value; } };
+    var queueEntry = function (v) {
+        f.value = v;
+        $schedule(thisGoroutine);
+    };
+    chan.$recvQueue.push(queueEntry);
+    $block();
+    return f;
+};
+var $close = function (chan) {
+    if (chan.$closed) {
+        $throwRuntimeError("close of closed channel");
+    }
+    chan.$closed = true;
+    while (true) {
+        var queuedSend = chan.$sendQueue.shift();
+        if (queuedSend === undefined) {
+            break;
+        }
+        queuedSend(true); /* will panic */
+    }
+    while (true) {
+        var queuedRecv = chan.$recvQueue.shift();
+        if (queuedRecv === undefined) {
+            break;
+        }
+        queuedRecv([chan.$elem.zero(), false]);
+    }
+};
+var $select = function (comms) {
+    var ready = [];
+    var selection = -1;
+    for (var i = 0; i < comms.length; i++) {
+        var comm = comms[i];
+        var chan = comm[0];
+        switch (comm.length) {
+            case 0: /* default */
+                selection = i;
+                break;
+            case 1: /* recv */
+                if (chan.$sendQueue.length !== 0 || chan.$buffer.length !== 0 || chan.$closed) {
+                    ready.push(i);
+                }
+                break;
+            case 2: /* send */
+                if (chan.$closed) {
+                    $throwRuntimeError("send on closed channel");
+                }
+                if (chan.$recvQueue.length !== 0 || chan.$buffer.length < chan.$capacity) {
+                    ready.push(i);
+                }
+                break;
+        }
+    }
+
+    if (ready.length !== 0) {
+        selection = ready[Math.floor(Math.random() * ready.length)];
+    }
+    if (selection !== -1) {
+        var comm = comms[selection];
+        switch (comm.length) {
+            case 0: /* default */
+                return [selection];
+            case 1: /* recv */
+                return [selection, $recv(comm[0])];
+            case 2: /* send */
+                $send(comm[0], comm[1]);
+                return [selection];
+        }
+    }
+
+    var entries = [];
+    var thisGoroutine = $curGoroutine;
+    var f = { $blk: function () { return this.selection; } };
+    var removeFromQueues = function () {
+        for (var i = 0; i < entries.length; i++) {
+            var entry = entries[i];
+            var queue = entry[0];
+            var index = queue.indexOf(entry[1]);
+            if (index !== -1) {
+                queue.splice(index, 1);
+            }
+        }
+    };
+    for (var i = 0; i < comms.length; i++) {
+        (function (i) {
+            var comm = comms[i];
+            switch (comm.length) {
+                case 1: /* recv */
+                    var queueEntry = function (value) {
+                        f.selection = [i, value];
+                        removeFromQueues();
+                        $schedule(thisGoroutine);
+                    };
+                    entries.push([comm[0].$recvQueue, queueEntry]);
+                    comm[0].$recvQueue.push(queueEntry);
+                    break;
+                case 2: /* send */
+                    var queueEntry = function () {
+                        if (comm[0].$closed) {
+                            $throwRuntimeError("send on closed channel");
+                        }
+                        f.selection = [i];
+                        removeFromQueues();
+                        $schedule(thisGoroutine);
+                        return comm[1];
+                    };
+                    entries.push([comm[0].$sendQueue, queueEntry]);
+                    comm[0].$sendQueue.push(queueEntry);
+                    break;
+            }
+        })(i);
+    }
+    $block();
+    return f;
+};
