+++ compiler/prelude/numeric.go
@@ -1,216 +1,8 @@
 package prelude
 
-const numeric = `
-var $min = Math.min;
-var $mod = function(x, y) { return x % y; };
-var $parseInt = parseInt;
-var $parseFloat = function(f) {
-  if (f !== undefined && f !== null && f.constructor === Number) {
-    return f;
-  }
-  return parseFloat(f);
-};
+import (
+	_ "embed"
+)
 
-var $froundBuf = new Float32Array(1);
-var $fround = Math.fround || function(f) {
-  $froundBuf[0] = f;
-  return $froundBuf[0];
-};
-
-var $imul = Math.imul || function(a, b) {
-  var ah = (a >>> 16) & 0xffff;
-  var al = a & 0xffff;
-  var bh = (b >>> 16) & 0xffff;
-  var bl = b & 0xffff;
-  return ((al * bl) + (((ah * bl + al * bh) << 16) >>> 0) >> 0);
-};
-
-var $floatKey = function(f) {
-  if (f !== f) {
-    $idCounter++;
-    return "NaN$" + $idCounter;
-  }
-  return String(f);
-};
-
-var $flatten64 = function(x) {
-  return x.$high * 4294967296 + x.$low;
-};
-
-var $shiftLeft64 = function(x, y) {
-  if (y === 0) {
-    return x;
-  }
-  if (y < 32) {
-    return new x.constructor(x.$high << y | x.$low >>> (32 - y), (x.$low << y) >>> 0);
-  }
-  if (y < 64) {
-    return new x.constructor(x.$low << (y - 32), 0);
-  }
-  return new x.constructor(0, 0);
-};
-
-var $shiftRightInt64 = function(x, y) {
-  if (y === 0) {
-    return x;
-  }
-  if (y < 32) {
-    return new x.constructor(x.$high >> y, (x.$low >>> y | x.$high << (32 - y)) >>> 0);
-  }
-  if (y < 64) {
-    return new x.constructor(x.$high >> 31, (x.$high >> (y - 32)) >>> 0);
-  }
-  if (x.$high < 0) {
-    return new x.constructor(-1, 4294967295);
-  }
-  return new x.constructor(0, 0);
-};
-
-var $shiftRightUint64 = function(x, y) {
-  if (y === 0) {
-    return x;
-  }
-  if (y < 32) {
-    return new x.constructor(x.$high >>> y, (x.$low >>> y | x.$high << (32 - y)) >>> 0);
-  }
-  if (y < 64) {
-    return new x.constructor(0, x.$high >>> (y - 32));
-  }
-  return new x.constructor(0, 0);
-};
-
-var $mul64 = function(x, y) {
-  var x48 = x.$high >>> 16;
-  var x32 = x.$high & 0xFFFF;
-  var x16 = x.$low >>> 16;
-  var x00 = x.$low & 0xFFFF;
-
-  var y48 = y.$high >>> 16;
-  var y32 = y.$high & 0xFFFF;
-  var y16 = y.$low >>> 16;
-  var y00 = y.$low & 0xFFFF;
-
-  var z48 = 0, z32 = 0, z16 = 0, z00 = 0;
-  z00 += x00 * y00;
-  z16 += z00 >>> 16;
-  z00 &= 0xFFFF;
-  z16 += x16 * y00;
-  z32 += z16 >>> 16;
-  z16 &= 0xFFFF;
-  z16 += x00 * y16;
-  z32 += z16 >>> 16;
-  z16 &= 0xFFFF;
-  z32 += x32 * y00;
-  z48 += z32 >>> 16;
-  z32 &= 0xFFFF;
-  z32 += x16 * y16;
-  z48 += z32 >>> 16;
-  z32 &= 0xFFFF;
-  z32 += x00 * y32;
-  z48 += z32 >>> 16;
-  z32 &= 0xFFFF;
-  z48 += x48 * y00 + x32 * y16 + x16 * y32 + x00 * y48;
-  z48 &= 0xFFFF;
-
-  var hi = ((z48 << 16) | z32) >>> 0;
-  var lo = ((z16 << 16) | z00) >>> 0;
-
-  var r = new x.constructor(hi, lo);
-  return r;
-};
-
-var $div64 = function(x, y, returnRemainder) {
-  if (y.$high === 0 && y.$low === 0) {
-    $throwRuntimeError("integer divide by zero");
-  }
-
-  var s = 1;
-  var rs = 1;
-
-  var xHigh = x.$high;
-  var xLow = x.$low;
-  if (xHigh < 0) {
-    s = -1;
-    rs = -1;
-    xHigh = -xHigh;
-    if (xLow !== 0) {
-      xHigh--;
-      xLow = 4294967296 - xLow;
-    }
-  }
-
-  var yHigh = y.$high;
-  var yLow = y.$low;
-  if (y.$high < 0) {
-    s *= -1;
-    yHigh = -yHigh;
-    if (yLow !== 0) {
-      yHigh--;
-      yLow = 4294967296 - yLow;
-    }
-  }
-
-  var high = 0, low = 0, n = 0;
-  while (yHigh < 2147483648 && ((xHigh > yHigh) || (xHigh === yHigh && xLow > yLow))) {
-    yHigh = (yHigh << 1 | yLow >>> 31) >>> 0;
-    yLow = (yLow << 1) >>> 0;
-    n++;
-  }
-  for (var i = 0; i <= n; i++) {
-    high = high << 1 | low >>> 31;
-    low = (low << 1) >>> 0;
-    if ((xHigh > yHigh) || (xHigh === yHigh && xLow >= yLow)) {
-      xHigh = xHigh - yHigh;
-      xLow = xLow - yLow;
-      if (xLow < 0) {
-        xHigh--;
-        xLow += 4294967296;
-      }
-      low++;
-      if (low === 4294967296) {
-        high++;
-        low = 0;
-      }
-    }
-    yLow = (yLow >>> 1 | yHigh << (32 - 1)) >>> 0;
-    yHigh = yHigh >>> 1;
-  }
-
-  if (returnRemainder) {
-    return new x.constructor(xHigh * rs, xLow * rs);
-  }
-  return new x.constructor(high * s, low * s);
-};
-
-var $divComplex = function(n, d) {
-  var ninf = n.$real === Infinity || n.$real === -Infinity || n.$imag === Infinity || n.$imag === -Infinity;
-  var dinf = d.$real === Infinity || d.$real === -Infinity || d.$imag === Infinity || d.$imag === -Infinity;
-  var nnan = !ninf && (n.$real !== n.$real || n.$imag !== n.$imag);
-  var dnan = !dinf && (d.$real !== d.$real || d.$imag !== d.$imag);
-  if(nnan || dnan) {
-    return new n.constructor(NaN, NaN);
-  }
-  if (ninf && !dinf) {
-    return new n.constructor(Infinity, Infinity);
-  }
-  if (!ninf && dinf) {
-    return new n.constructor(0, 0);
-  }
-  if (d.$real === 0 && d.$imag === 0) {
-    if (n.$real === 0 && n.$imag === 0) {
-      return new n.constructor(NaN, NaN);
-    }
-    return new n.constructor(Infinity, Infinity);
-  }
-  var a = Math.abs(d.$real);
-  var b = Math.abs(d.$imag);
-  if (a <= b) {
-    var ratio = d.$real / d.$imag;
-    var denom = d.$real * ratio + d.$imag;
-    return new n.constructor((n.$real * ratio + n.$imag) / denom, (n.$imag * ratio - n.$real) / denom);
-  }
-  var ratio = d.$imag / d.$real;
-  var denom = d.$imag * ratio + d.$real;
-  return new n.constructor((n.$imag * ratio + n.$real) / denom, (n.$imag - n.$real * ratio) / denom);
-};
-`
+//go:embed numeric.js
+var numeric string
