+++ compiler/prelude/types.go
@@ -1,774 +1,8 @@
 package prelude
 
-const types = `
-var $kindBool = 1;
-var $kindInt = 2;
-var $kindInt8 = 3;
-var $kindInt16 = 4;
-var $kindInt32 = 5;
-var $kindInt64 = 6;
-var $kindUint = 7;
-var $kindUint8 = 8;
-var $kindUint16 = 9;
-var $kindUint32 = 10;
-var $kindUint64 = 11;
-var $kindUintptr = 12;
-var $kindFloat32 = 13;
-var $kindFloat64 = 14;
-var $kindComplex64 = 15;
-var $kindComplex128 = 16;
-var $kindArray = 17;
-var $kindChan = 18;
-var $kindFunc = 19;
-var $kindInterface = 20;
-var $kindMap = 21;
-var $kindPtr = 22;
-var $kindSlice = 23;
-var $kindString = 24;
-var $kindStruct = 25;
-var $kindUnsafePointer = 26;
+import (
+	_ "embed"
+)
 
-var $methodSynthesizers = [];
-var $addMethodSynthesizer = function(f) {
-  if ($methodSynthesizers === null) {
-    f();
-    return;
-  }
-  $methodSynthesizers.push(f);
-};
-var $synthesizeMethods = function() {
-  $methodSynthesizers.forEach(function(f) { f(); });
-  $methodSynthesizers = null;
-};
-
-var $ifaceKeyFor = function(x) {
-  if (x === $ifaceNil) {
-    return 'nil';
-  }
-  var c = x.constructor;
-  return c.string + '$' + c.keyFor(x.$val);
-};
-
-var $identity = function(x) { return x; };
-
-var $typeIDCounter = 0;
-
-var $idKey = function(x) {
-  if (x.$id === undefined) {
-    $idCounter++;
-    x.$id = $idCounter;
-  }
-  return String(x.$id);
-};
-
-// Creates constructor functions for array pointer types. Returns a new function
-// instace each time to make sure each type is independent of the other.
-var $arrayPtrCtor = function() {
-  return function(array) {
-    this.$get = function() { return array; };
-    this.$set = function(v) { typ.copy(this, v); };
-    this.$val = array;
-  }
-}
-
-var $newType = function(size, kind, string, named, pkg, exported, constructor) {
-  var typ;
-  switch(kind) {
-  case $kindBool:
-  case $kindInt:
-  case $kindInt8:
-  case $kindInt16:
-  case $kindInt32:
-  case $kindUint:
-  case $kindUint8:
-  case $kindUint16:
-  case $kindUint32:
-  case $kindUintptr:
-  case $kindUnsafePointer:
-    typ = function(v) { this.$val = v; };
-    typ.wrapped = true;
-    typ.keyFor = $identity;
-    break;
-
-  case $kindString:
-    typ = function(v) { this.$val = v; };
-    typ.wrapped = true;
-    typ.keyFor = function(x) { return "$" + x; };
-    break;
-
-  case $kindFloat32:
-  case $kindFloat64:
-    typ = function(v) { this.$val = v; };
-    typ.wrapped = true;
-    typ.keyFor = function(x) { return $floatKey(x); };
-    break;
-
-  case $kindInt64:
-    typ = function(high, low) {
-      this.$high = (high + Math.floor(Math.ceil(low) / 4294967296)) >> 0;
-      this.$low = low >>> 0;
-      this.$val = this;
-    };
-    typ.keyFor = function(x) { return x.$high + "$" + x.$low; };
-    break;
-
-  case $kindUint64:
-    typ = function(high, low) {
-      this.$high = (high + Math.floor(Math.ceil(low) / 4294967296)) >>> 0;
-      this.$low = low >>> 0;
-      this.$val = this;
-    };
-    typ.keyFor = function(x) { return x.$high + "$" + x.$low; };
-    break;
-
-  case $kindComplex64:
-    typ = function(real, imag) {
-      this.$real = $fround(real);
-      this.$imag = $fround(imag);
-      this.$val = this;
-    };
-    typ.keyFor = function(x) { return x.$real + "$" + x.$imag; };
-    break;
-
-  case $kindComplex128:
-    typ = function(real, imag) {
-      this.$real = real;
-      this.$imag = imag;
-      this.$val = this;
-    };
-    typ.keyFor = function(x) { return x.$real + "$" + x.$imag; };
-    break;
-
-  case $kindArray:
-    typ = function(v) { this.$val = v; };
-    typ.wrapped = true;
-    typ.ptr = $newType(4, $kindPtr, "*" + string, false, "", false, $arrayPtrCtor());
-    typ.init = function(elem, len) {
-      typ.elem = elem;
-      typ.len = len;
-      typ.comparable = elem.comparable;
-      typ.keyFor = function(x) {
-        return Array.prototype.join.call($mapArray(x, function(e) {
-          return String(elem.keyFor(e)).replace(/\\/g, "\\\\").replace(/\$/g, "\\$");
-        }), "$");
-      };
-      typ.copy = function(dst, src) {
-        $copyArray(dst, src, 0, 0, src.length, elem);
-      };
-      typ.ptr.init(typ);
-      Object.defineProperty(typ.ptr.nil, "nilCheck", { get: $throwNilPointerError });
-    };
-    break;
-
-  case $kindChan:
-    typ = function(v) { this.$val = v; };
-    typ.wrapped = true;
-    typ.keyFor = $idKey;
-    typ.init = function(elem, sendOnly, recvOnly) {
-      typ.elem = elem;
-      typ.sendOnly = sendOnly;
-      typ.recvOnly = recvOnly;
-    };
-    break;
-
-  case $kindFunc:
-    typ = function(v) { this.$val = v; };
-    typ.wrapped = true;
-    typ.init = function(params, results, variadic) {
-      typ.params = params;
-      typ.results = results;
-      typ.variadic = variadic;
-      typ.comparable = false;
-    };
-    break;
-
-  case $kindInterface:
-    typ = { implementedBy: {}, missingMethodFor: {} };
-    typ.keyFor = $ifaceKeyFor;
-    typ.init = function(methods) {
-      typ.methods = methods;
-      methods.forEach(function(m) {
-        $ifaceNil[m.prop] = $throwNilPointerError;
-      });
-    };
-    break;
-
-  case $kindMap:
-    typ = function(v) { this.$val = v; };
-    typ.wrapped = true;
-    typ.init = function(key, elem) {
-      typ.key = key;
-      typ.elem = elem;
-      typ.comparable = false;
-    };
-    break;
-
-  case $kindPtr:
-    typ = constructor || function(getter, setter, target) {
-      this.$get = getter;
-      this.$set = setter;
-      this.$target = target;
-      this.$val = this;
-    };
-    typ.keyFor = $idKey;
-    typ.init = function(elem) {
-      typ.elem = elem;
-      typ.wrapped = (elem.kind === $kindArray);
-      typ.nil = new typ($throwNilPointerError, $throwNilPointerError);
-    };
-    break;
-
-  case $kindSlice:
-    typ = function(array) {
-      if (array.constructor !== typ.nativeArray) {
-        array = new typ.nativeArray(array);
-      }
-      this.$array = array;
-      this.$offset = 0;
-      this.$length = array.length;
-      this.$capacity = array.length;
-      this.$val = this;
-    };
-    typ.init = function(elem) {
-      typ.elem = elem;
-      typ.comparable = false;
-      typ.nativeArray = $nativeArray(elem.kind);
-      typ.nil = new typ([]);
-    };
-    break;
-
-  case $kindStruct:
-    typ = function(v) { this.$val = v; };
-    typ.wrapped = true;
-    typ.ptr = $newType(4, $kindPtr, "*" + string, false, pkg, exported, constructor);
-    typ.ptr.elem = typ;
-    typ.ptr.prototype.$get = function() { return this; };
-    typ.ptr.prototype.$set = function(v) { typ.copy(this, v); };
-    typ.init = function(pkgPath, fields) {
-      typ.pkgPath = pkgPath;
-      typ.fields = fields;
-      fields.forEach(function(f) {
-        if (!f.typ.comparable) {
-          typ.comparable = false;
-        }
-      });
-      typ.keyFor = function(x) {
-        var val = x.$val;
-        return $mapArray(fields, function(f) {
-          return String(f.typ.keyFor(val[f.prop])).replace(/\\/g, "\\\\").replace(/\$/g, "\\$");
-        }).join("$");
-      };
-      typ.copy = function(dst, src) {
-        for (var i = 0; i < fields.length; i++) {
-          var f = fields[i];
-          switch (f.typ.kind) {
-          case $kindArray:
-          case $kindStruct:
-            f.typ.copy(dst[f.prop], src[f.prop]);
-            continue;
-          default:
-            dst[f.prop] = src[f.prop];
-            continue;
-          }
-        }
-      };
-      /* nil value */
-      var properties = {};
-      fields.forEach(function(f) {
-        properties[f.prop] = { get: $throwNilPointerError, set: $throwNilPointerError };
-      });
-      typ.ptr.nil = Object.create(constructor.prototype, properties);
-      typ.ptr.nil.$val = typ.ptr.nil;
-      /* methods for embedded fields */
-      $addMethodSynthesizer(function() {
-        var synthesizeMethod = function(target, m, f) {
-          if (target.prototype[m.prop] !== undefined) { return; }
-          target.prototype[m.prop] = function() {
-            var v = this.$val[f.prop];
-            if (f.typ === $jsObjectPtr) {
-              v = new $jsObjectPtr(v);
-            }
-            if (v.$val === undefined) {
-              v = new f.typ(v);
-            }
-            return v[m.prop].apply(v, arguments);
-          };
-        };
-        fields.forEach(function(f) {
-          if (f.embedded) {
-            $methodSet(f.typ).forEach(function(m) {
-              synthesizeMethod(typ, m, f);
-              synthesizeMethod(typ.ptr, m, f);
-            });
-            $methodSet($ptrType(f.typ)).forEach(function(m) {
-              synthesizeMethod(typ.ptr, m, f);
-            });
-          }
-        });
-      });
-    };
-    break;
-
-  default:
-    $panic(new $String("invalid kind: " + kind));
-  }
-
-  switch (kind) {
-  case $kindBool:
-  case $kindMap:
-    typ.zero = function() { return false; };
-    break;
-
-  case $kindInt:
-  case $kindInt8:
-  case $kindInt16:
-  case $kindInt32:
-  case $kindUint:
-  case $kindUint8 :
-  case $kindUint16:
-  case $kindUint32:
-  case $kindUintptr:
-  case $kindUnsafePointer:
-  case $kindFloat32:
-  case $kindFloat64:
-    typ.zero = function() { return 0; };
-    break;
-
-  case $kindString:
-    typ.zero = function() { return ""; };
-    break;
-
-  case $kindInt64:
-  case $kindUint64:
-  case $kindComplex64:
-  case $kindComplex128:
-    var zero = new typ(0, 0);
-    typ.zero = function() { return zero; };
-    break;
-
-  case $kindPtr:
-  case $kindSlice:
-    typ.zero = function() { return typ.nil; };
-    break;
-
-  case $kindChan:
-    typ.zero = function() { return $chanNil; };
-    break;
-
-  case $kindFunc:
-    typ.zero = function() { return $throwNilPointerError; };
-    break;
-
-  case $kindInterface:
-    typ.zero = function() { return $ifaceNil; };
-    break;
-
-  case $kindArray:
-    typ.zero = function() {
-      var arrayClass = $nativeArray(typ.elem.kind);
-      if (arrayClass !== Array) {
-        return new arrayClass(typ.len);
-      }
-      var array = new Array(typ.len);
-      for (var i = 0; i < typ.len; i++) {
-        array[i] = typ.elem.zero();
-      }
-      return array;
-    };
-    break;
-
-  case $kindStruct:
-    typ.zero = function() { return new typ.ptr(); };
-    break;
-
-  default:
-    $panic(new $String("invalid kind: " + kind));
-  }
-
-  typ.id = $typeIDCounter;
-  $typeIDCounter++;
-  typ.size = size;
-  typ.kind = kind;
-  typ.string = string;
-  typ.named = named;
-  typ.pkg = pkg;
-  typ.exported = exported;
-  typ.methods = [];
-  typ.methodSetCache = null;
-  typ.comparable = true;
-  return typ;
-};
-
-var $methodSet = function(typ) {
-  if (typ.methodSetCache !== null) {
-    return typ.methodSetCache;
-  }
-  var base = {};
-
-  var isPtr = (typ.kind === $kindPtr);
-  if (isPtr && typ.elem.kind === $kindInterface) {
-    typ.methodSetCache = [];
-    return [];
-  }
-
-  var current = [{typ: isPtr ? typ.elem : typ, indirect: isPtr}];
-
-  var seen = {};
-
-  while (current.length > 0) {
-    var next = [];
-    var mset = [];
-
-    current.forEach(function(e) {
-      if (seen[e.typ.string]) {
-        return;
-      }
-      seen[e.typ.string] = true;
-
-      if (e.typ.named) {
-        mset = mset.concat(e.typ.methods);
-        if (e.indirect) {
-          mset = mset.concat($ptrType(e.typ).methods);
-        }
-      }
-
-      switch (e.typ.kind) {
-      case $kindStruct:
-        e.typ.fields.forEach(function(f) {
-          if (f.embedded) {
-            var fTyp = f.typ;
-            var fIsPtr = (fTyp.kind === $kindPtr);
-            next.push({typ: fIsPtr ? fTyp.elem : fTyp, indirect: e.indirect || fIsPtr});
-          }
-        });
-        break;
-
-      case $kindInterface:
-        mset = mset.concat(e.typ.methods);
-        break;
-      }
-    });
-
-    mset.forEach(function(m) {
-      if (base[m.name] === undefined) {
-        base[m.name] = m;
-      }
-    });
-
-    current = next;
-  }
-
-  typ.methodSetCache = [];
-  Object.keys(base).sort().forEach(function(name) {
-    typ.methodSetCache.push(base[name]);
-  });
-  return typ.methodSetCache;
-};
-
-var $Bool          = $newType( 1, $kindBool,          "bool",           true, "", false, null);
-var $Int           = $newType( 4, $kindInt,           "int",            true, "", false, null);
-var $Int8          = $newType( 1, $kindInt8,          "int8",           true, "", false, null);
-var $Int16         = $newType( 2, $kindInt16,         "int16",          true, "", false, null);
-var $Int32         = $newType( 4, $kindInt32,         "int32",          true, "", false, null);
-var $Int64         = $newType( 8, $kindInt64,         "int64",          true, "", false, null);
-var $Uint          = $newType( 4, $kindUint,          "uint",           true, "", false, null);
-var $Uint8         = $newType( 1, $kindUint8,         "uint8",          true, "", false, null);
-var $Uint16        = $newType( 2, $kindUint16,        "uint16",         true, "", false, null);
-var $Uint32        = $newType( 4, $kindUint32,        "uint32",         true, "", false, null);
-var $Uint64        = $newType( 8, $kindUint64,        "uint64",         true, "", false, null);
-var $Uintptr       = $newType( 4, $kindUintptr,       "uintptr",        true, "", false, null);
-var $Float32       = $newType( 4, $kindFloat32,       "float32",        true, "", false, null);
-var $Float64       = $newType( 8, $kindFloat64,       "float64",        true, "", false, null);
-var $Complex64     = $newType( 8, $kindComplex64,     "complex64",      true, "", false, null);
-var $Complex128    = $newType(16, $kindComplex128,    "complex128",     true, "", false, null);
-var $String        = $newType( 8, $kindString,        "string",         true, "", false, null);
-var $UnsafePointer = $newType( 4, $kindUnsafePointer, "unsafe.Pointer", true, "unsafe", false, null);
-
-var $nativeArray = function(elemKind) {
-  switch (elemKind) {
-  case $kindInt:
-    return Int32Array;
-  case $kindInt8:
-    return Int8Array;
-  case $kindInt16:
-    return Int16Array;
-  case $kindInt32:
-    return Int32Array;
-  case $kindUint:
-    return Uint32Array;
-  case $kindUint8:
-    return Uint8Array;
-  case $kindUint16:
-    return Uint16Array;
-  case $kindUint32:
-    return Uint32Array;
-  case $kindUintptr:
-    return Uint32Array;
-  case $kindFloat32:
-    return Float32Array;
-  case $kindFloat64:
-    return Float64Array;
-  default:
-    return Array;
-  }
-};
-var $toNativeArray = function(elemKind, array) {
-  var nativeArray = $nativeArray(elemKind);
-  if (nativeArray === Array) {
-    return array;
-  }
-  return new nativeArray(array);
-};
-var $arrayTypes = {};
-var $arrayType = function(elem, len) {
-  var typeKey = elem.id + "$" + len;
-  var typ = $arrayTypes[typeKey];
-  if (typ === undefined) {
-    typ = $newType(elem.size*len, $kindArray, "[" + len + "]" + elem.string, false, "", false, null);
-    $arrayTypes[typeKey] = typ;
-    typ.init(elem, len);
-  }
-  return typ;
-};
-
-var $chanType = function(elem, sendOnly, recvOnly) {
-  var string = (recvOnly ? "<-" : "") + "chan" + (sendOnly ? "<- " : " ");
-  if (!sendOnly && !recvOnly && (elem.string[0] == "<")) {
-    string += "(" + elem.string + ")";
-  } else {
-    string += elem.string;
-  }
-  var field = sendOnly ? "SendChan" : (recvOnly ? "RecvChan" : "Chan");
-  var typ = elem[field];
-  if (typ === undefined) {
-    typ = $newType(4, $kindChan, string, false, "", false, null);
-    elem[field] = typ;
-    typ.init(elem, sendOnly, recvOnly);
-  }
-  return typ;
-};
-var $Chan = function(elem, capacity) {
-  if (capacity < 0 || capacity > 2147483647) {
-    $throwRuntimeError("makechan: size out of range");
-  }
-  this.$elem = elem;
-  this.$capacity = capacity;
-  this.$buffer = [];
-  this.$sendQueue = [];
-  this.$recvQueue = [];
-  this.$closed = false;
-};
-var $chanNil = new $Chan(null, 0);
-$chanNil.$sendQueue = $chanNil.$recvQueue = { length: 0, push: function() {}, shift: function() { return undefined; }, indexOf: function() { return -1; } };
-
-var $funcTypes = {};
-var $funcType = function(params, results, variadic) {
-  var typeKey = $mapArray(params, function(p) { return p.id; }).join(",") + "$" + $mapArray(results, function(r) { return r.id; }).join(",") + "$" + variadic;
-  var typ = $funcTypes[typeKey];
-  if (typ === undefined) {
-    var paramTypes = $mapArray(params, function(p) { return p.string; });
-    if (variadic) {
-      paramTypes[paramTypes.length - 1] = "..." + paramTypes[paramTypes.length - 1].substr(2);
-    }
-    var string = "func(" + paramTypes.join(", ") + ")";
-    if (results.length === 1) {
-      string += " " + results[0].string;
-    } else if (results.length > 1) {
-      string += " (" + $mapArray(results, function(r) { return r.string; }).join(", ") + ")";
-    }
-    typ = $newType(4, $kindFunc, string, false, "", false, null);
-    $funcTypes[typeKey] = typ;
-    typ.init(params, results, variadic);
-  }
-  return typ;
-};
-
-var $interfaceTypes = {};
-var $interfaceType = function(methods) {
-  var typeKey = $mapArray(methods, function(m) { return m.pkg + "," + m.name + "," + m.typ.id; }).join("$");
-  var typ = $interfaceTypes[typeKey];
-  if (typ === undefined) {
-    var string = "interface {}";
-    if (methods.length !== 0) {
-      string = "interface { " + $mapArray(methods, function(m) {
-        return (m.pkg !== "" ? m.pkg + "." : "") + m.name + m.typ.string.substr(4);
-      }).join("; ") + " }";
-    }
-    typ = $newType(8, $kindInterface, string, false, "", false, null);
-    $interfaceTypes[typeKey] = typ;
-    typ.init(methods);
-  }
-  return typ;
-};
-var $emptyInterface = $interfaceType([]);
-var $ifaceNil = {};
-var $error = $newType(8, $kindInterface, "error", true, "", false, null);
-$error.init([{prop: "Error", name: "Error", pkg: "", typ: $funcType([], [$String], false)}]);
-
-var $mapTypes = {};
-var $mapType = function(key, elem) {
-  var typeKey = key.id + "$" + elem.id;
-  var typ = $mapTypes[typeKey];
-  if (typ === undefined) {
-    typ = $newType(4, $kindMap, "map[" + key.string + "]" + elem.string, false, "", false, null);
-    $mapTypes[typeKey] = typ;
-    typ.init(key, elem);
-  }
-  return typ;
-};
-var $makeMap = function(keyForFunc, entries) {
-  var m = new Map();
-  for (var i = 0; i < entries.length; i++) {
-    var e = entries[i];
-    m.set(keyForFunc(e.k), e);
-  }
-  return m;
-};
-
-var $ptrType = function(elem) {
-  var typ = elem.ptr;
-  if (typ === undefined) {
-    typ = $newType(4, $kindPtr, "*" + elem.string, false, "", elem.exported, null);
-    elem.ptr = typ;
-    typ.init(elem);
-  }
-  return typ;
-};
-
-var $newDataPointer = function(data, constructor) {
-  if (constructor.elem.kind === $kindStruct) {
-    return data;
-  }
-  return new constructor(function() { return data; }, function(v) { data = v; });
-};
-
-var $indexPtr = function(array, index, constructor) {
-  if (array.buffer) {
-    // Pointers to the same underlying ArrayBuffer share cache.
-    var cache = array.buffer.$ptr = array.buffer.$ptr || {};
-    // Pointers of different primitive types are non-comparable and stored in different caches.
-    var typeCache = cache[array.name] = cache[array.name] || {};
-    var cacheIdx = array.BYTES_PER_ELEMENT * index + array.byteOffset;
-    return typeCache[cacheIdx] || (typeCache[cacheIdx] = new constructor(function() { return array[index]; }, function(v) { array[index] = v; }));
-  } else {
-    array.$ptr = array.$ptr || {};
-    return array.$ptr[index] || (array.$ptr[index] = new constructor(function() { return array[index]; }, function(v) { array[index] = v; }));
-  }
-};
-
-var $sliceType = function(elem) {
-  var typ = elem.slice;
-  if (typ === undefined) {
-    typ = $newType(12, $kindSlice, "[]" + elem.string, false, "", false, null);
-    elem.slice = typ;
-    typ.init(elem);
-  }
-  return typ;
-};
-var $makeSlice = function(typ, length, capacity) {
-  capacity = capacity || length;
-  if (length < 0 || length > 2147483647) {
-    $throwRuntimeError("makeslice: len out of range");
-  }
-  if (capacity < 0 || capacity < length || capacity > 2147483647) {
-    $throwRuntimeError("makeslice: cap out of range");
-  }
-  var array = new typ.nativeArray(capacity);
-  if (typ.nativeArray === Array) {
-    for (var i = 0; i < capacity; i++) {
-      array[i] = typ.elem.zero();
-    }
-  }
-  var slice = new typ(array);
-  slice.$length = length;
-  return slice;
-};
-
-var $structTypes = {};
-var $structType = function(pkgPath, fields) {
-  var typeKey = $mapArray(fields, function(f) { return f.name + "," + f.typ.id + "," + f.tag; }).join("$");
-  var typ = $structTypes[typeKey];
-  if (typ === undefined) {
-    var string = "struct { " + $mapArray(fields, function(f) {
-      var str = f.typ.string + (f.tag !== "" ? (" \"" + f.tag.replace(/\\/g, "\\\\").replace(/"/g, "\\\"") + "\"") : "");
-      if (f.embedded) {
-        return str;
-      }
-      return f.name + " " + str;
-    }).join("; ") + " }";
-    if (fields.length === 0) {
-      string = "struct {}";
-    }
-    typ = $newType(0, $kindStruct, string, false, "", false, function() {
-      this.$val = this;
-      for (var i = 0; i < fields.length; i++) {
-        var f = fields[i];
-        if (f.name == '_') {
-          continue;
-        }
-        var arg = arguments[i];
-        this[f.prop] = arg !== undefined ? arg : f.typ.zero();
-      }
-    });
-    $structTypes[typeKey] = typ;
-    typ.init(pkgPath, fields);
-  }
-  return typ;
-};
-
-var $assertType = function(value, type, returnTuple) {
-  var isInterface = (type.kind === $kindInterface), ok, missingMethod = "";
-  if (value === $ifaceNil) {
-    ok = false;
-  } else if (!isInterface) {
-    ok = value.constructor === type;
-  } else {
-    var valueTypeString = value.constructor.string;
-    ok = type.implementedBy[valueTypeString];
-    if (ok === undefined) {
-      ok = true;
-      var valueMethodSet = $methodSet(value.constructor);
-      var interfaceMethods = type.methods;
-      for (var i = 0; i < interfaceMethods.length; i++) {
-        var tm = interfaceMethods[i];
-        var found = false;
-        for (var j = 0; j < valueMethodSet.length; j++) {
-          var vm = valueMethodSet[j];
-          if (vm.name === tm.name && vm.pkg === tm.pkg && vm.typ === tm.typ) {
-            found = true;
-            break;
-          }
-        }
-        if (!found) {
-          ok = false;
-          type.missingMethodFor[valueTypeString] = tm.name;
-          break;
-        }
-      }
-      type.implementedBy[valueTypeString] = ok;
-    }
-    if (!ok) {
-      missingMethod = type.missingMethodFor[valueTypeString];
-    }
-  }
-
-  if (!ok) {
-    if (returnTuple) {
-      return [type.zero(), false];
-    }
-    $panic(new $packages["runtime"].TypeAssertionError.ptr(
-      $packages["runtime"]._type.ptr.nil,
-      (value === $ifaceNil ? $packages["runtime"]._type.ptr.nil : new $packages["runtime"]._type.ptr(value.constructor.string)),
-      new $packages["runtime"]._type.ptr(type.string),
-      missingMethod));
-  }
-
-  if (!isInterface) {
-    value = value.$val;
-  }
-  if (type === $jsObjectPtr) {
-    value = value.object;
-  }
-  return returnTuple ? [value, true] : value;
-};
-`
+//go:embed types.js
+var types string
